Rename package `throttle` -> `limit`
- previous commit only changes directory name
